refactor: only move sw main, not app
