+++ apps/fulfillment/src/app.ts
@@ -3,12 +3,10 @@ import { injectEnv, PageMetaResolver } from '@spryker-oryx/core';
 import { ContentBackendUrl, experienceFeature } from '@spryker-oryx/experience';
 import { formFeature } from '@spryker-oryx/form';
 import { labsFeatures } from '@spryker-oryx/labs';
-import {
-  offlineFulfillmentFeatures,
-  fallbackEnv,
-} from '@spryker-oryx/presets/fulfillment';
+import { offlineFulfillmentFeatures } from '@spryker-oryx/presets/fulfillment';
 import { siteFeature } from '@spryker-oryx/site';
 import { fulfillmentTheme } from '@spryker-oryx/themes';
+import { fallbackEnv } from './fallback-env';
 
 const env = import.meta.env;
