+++ apps/fulfillment/sw/app.ts
@@ -1,8 +1,6 @@
 import { appBuilder } from '@spryker-oryx/application';
-import {
-  fallbackEnv,
-  offlineServiceWorkerFulfillmentFeatures,
-} from '@spryker-oryx/presets/fulfillment';
+import { offlineServiceWorkerFulfillmentFeatures } from '@spryker-oryx/presets/fulfillment';
+import { fallbackEnv } from '../src/fallback-env';
 
 export const app = appBuilder()
   .withEnvironment({ ...fallbackEnv, ...process.env })
