+++ apps/fulfillment/sw/main.ts
@@ -1,1 +1,6 @@
-export * from '@spryker-oryx/application/sw';
+import '@spryker-oryx/application/sw';
+import { app } from './app';
+
+app
+  .then(() => console.debug('Service worker app started!'))
+  .catch(console.error);
