+++ libs/template/presets/fulfillment/index.ts
@@ -1,5 +1,4 @@
 export * from './app';
 export * from './components';
-export * from './fallback-env';
 export * from './feature';
 export * from './offline';
